requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.11.0
cloudinary==1.43.0
astunparse==1.6.3
attrs==25.4.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comtypes==1.4.13
contourpy==1.3.3
cycler==0.12.1
python-dotenv==1.2.1
et_xmlfile==2.0.0
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.10.0
fuzzywuzzy==0.18.0
gast==0.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.72.0
greenlet==3.3.2
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
jax==0.8.0
jaxlib==0.8.0
Jinja2==3.1.6
joblib==1.5.2
keras==3.12.0
keyboard==0.13.5
kiwisolver==1.4.9
Levenshtein==0.27.3
libclang==18.1.1
llvmlite==0.45.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mediapipe==0.10.14
ml_dtypes==0.5.3
more-itertools==10.8.0
MouseInfo==0.1.3
mpmath==1.3.0
mss==10.1.0
namex==0.1.0
networkx==3.5
numba==0.62.1
numpy==2.2.6
openai-whisper==20250625
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.18.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
playwright==1.58.0
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycaw==20251023
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydub==0.25.1
pyee==13.0.1
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==2.0.1
pyparsing==3.2.5
pyperclip==1.11.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-Levenshtein==0.27.3
python-vlc==3.0.21203
pyttsx3==2.99
pytweening==1.2.0
pytz==2025.2
pywin32==311
RapidFuzz==3.14.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.3
selenium==4.38.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.3
SpeechRecognition==3.14.3
srt==3.5.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
threadpoolctl==3.6.0
tiktoken==0.12.0
torch==2.9.1
tqdm==4.67.1
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vosk==0.3.45
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.1
wsproto==1.3.1
youtube-search-python==1.6.6
